==> 1. Data Pre-processing.ipynb <==
## CREDIT CARD DEFAULT PREDICTION AND MODEL INTERPRETATION                             USING POST- HOC EXPLAINABILITY TECHNIQUES 

### Synopysis

Objective – Predict the probability of a customer defaulting payment for the credit card the subsequent month, based on past information. The past information is provided in the dataset. This probability will help the collections team to prioritise follow up with customers who have a high propensity of defaulting. The goal of the research is to explore and exhibit the capabilities of various model agnostic interpretability techniques to generate explainable insights into prediction model outcomes

1. Getting the Data - https://archive.ics.uci.edu/ml/datasets/default+of+credit+card+clients
2. Data Quality checks


* Input : Raw dataset from UCI (default_of_credit_card_clients.xls)
* Outcome : Cleaned dataset (inpt_df_processed.csv)

### Attributes description
This study uses 23 variables as explanatory variables, extracted/interpreted from : 
    
```
   ----------------------------------------------------------------------------------------
   Feature Name         Feature Description        
   -------------------- -------------------------------------------------------------------
   limit_bal            Amount of the given credit (NT dollar): it includes both the individual 
                        consumer credit and his/her family (supplementary) credit.
                        
   sex                  Gender (1 = male; 2 = female)
   
   education            Education (1 = graduate school; 2 = university; 3 = high school; 4 = others)
   
   marriage             Marital status (1 = married; 2 = single; 3 = others)
   
   age                  Age (in years)
   
   pay_1 - pay_6        History of past payment. Past monthly payment records 
                        From April to September, 2005 as follows:                     
                        pay_1 = the repayment status in September, 2005
                        pay_2 = the repayment status in August, 2005
                        ...
                        pay_6 = the repayment status in April, 2005 
                        
                        The measurement scale for the repayment status is: 
                        -1 = pay duly; 
                        1 = payment delay for one month 
                        2 = payment delay for two months
                        ...
                        8 = payment delay for eight months 
                        9 = payment delay for nine months and above
                        
   bill_amt1-bill_amt5  Amount of bill statement (NT dollar). 
                        bill_amt1 = amount of bill statement in September, 2005 
                        bill_amt2 = amount of bill statement in August, 2005
                        ...
                        bill_amt6= amount of bill statement in April, 2005 
                        
   pay_amt1-pay_amt6    Amount of previous payment (NT dollar)
                        pay_amt1 = amount paid in September, 2005
                        pay_amt2 = amount paid in August, 2005
                        ...
                        pay_amt6 = amount paid in April, 2005 
   ----------------------------------------------------------------------------------------
```

# Import libraries
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

# Import plotting libraries
import seaborn as sns
import matplotlib.pyplot as plt
import pylab

#### 2. Data Importing

#--------------------------------------------------------------
#Importing the excel file as python dataframe
#--------------------------------------------------------------

inpt_df=pd.read_excel('default_of_credit_card_clients.xls',header=1)
inpt_df.head(10)

#--------------------------------------------------------------
#Size and shape of the dataset - There are 30,000 observations and 25 columns
#--------------------------------------------------------------

print("Shape of data:", inpt_df.shape)

print("Numerical features count: ", len(inpt_df.dtypes[inpt_df.dtypes != "object"].index))

print("Categorical features count: ", len(inpt_df.dtypes[inpt_df.dtypes == "object"].index))

#### 2. Data Quality checks

#--------------------------------------------------------------
#1. Check for missing/null values - There are no missing values
#--------------------------------------------------------------

inpt_df.isnull().sum()

#--------------------------------------------------------------
#2. Check for duplicates - No duplicate rows found
#--------------------------------------------------------------

print("No of unique records: ", inpt_df['ID'].nunique())

#--------------------------------------------------------------
#3. Elimination of unnecessary variable and data Cleaning
#--------------------------------------------------------------

# Dropping unnecessary feature(ID) which do not contribute significantly
inpt_df = inpt_df.drop(["ID"],axis=1)

#Renaming Pay_0 to Pay_1 to correct the numbering order of payment status
inpt_df.rename(columns={'PAY_0':'PAY_1'}, inplace=True)
inpt_df.rename(columns={'default payment next month':'DEFAULT'}, inplace=True)

#--------------------------------------------------------------
#4. Data Correctness check 
#--------------------------------------------------------------

print(inpt_df.AGE[inpt_df.AGE<18].count(),
      inpt_df.LIMIT_BAL[inpt_df.LIMIT_BAL<0].count(),
      inpt_df.DEFAULT[inpt_df.DEFAULT>1].count())

#4.A Distribution of categorical variables: EDUCATION, MARRIAGE and SEX

    #There is no description available for 0 in Education and Marriage fields
    #EDUCATION: (1=graduate school, 2=university, 3=high school, 4=others, 5=unknown, 6=unknown)
    #MARRIAGE: (1=married, 2=single, 3=others) 

print(inpt_df['EDUCATION'].value_counts())
print(inpt_df['MARRIAGE'].value_counts())
print(inpt_df['SEX'].value_counts())

#4.B Replacing all unknown class values as one category 'Others'

inpt_df["EDUCATION"]=inpt_df["EDUCATION"].map({0:4,1:1,2:2,3:3,4:4,5:4,6:4})
inpt_df["MARRIAGE"]=inpt_df["MARRIAGE"].map({0:3,1:1,2:2,3:3})

#4.C Payment status for 6 months - There is no description for -2 and 0

    #As per dataset creator's response in https://www.kaggle.com/uciml/default-of-credit-card-clients-dataset/discussion/34608
    # -2 : No consumption/transaction; -1 : Paid in full; 0 : The use of revolving credit; 
    
inpt_df[['PAY_1', 'PAY_2', 'PAY_3','PAY_4','PAY_5','PAY_6']].apply(pd.Series.value_counts)

#4.D Standardizing the values of payment status field by marking -2,-1 and 0 as 0

repl = (inpt_df.PAY_1 == -2) | (inpt_df.PAY_1 == -1) | (inpt_df.PAY_1 == 0)
inpt_df.loc[repl, 'PAY_1'] = 0
repl = (inpt_df.PAY_2 == -2) | (inpt_df.PAY_2 == -1) | (inpt_df.PAY_2 == 0)
inpt_df.loc[repl, 'PAY_2'] = 0
repl = (inpt_df.PAY_3 == -2) | (inpt_df.PAY_3 == -1) | (inpt_df.PAY_3 == 0)
inpt_df.loc[repl, 'PAY_3'] = 0
repl = (inpt_df.PAY_4 == -2) | (inpt_df.PAY_4 == -1) | (inpt_df.PAY_4 == 0)
inpt_df.loc[repl, 'PAY_4'] = 0
repl = (inpt_df.PAY_5 == -2) | (inpt_df.PAY_5 == -1) | (inpt_df.PAY_5 == 0)
inpt_df.loc[repl, 'PAY_5'] = 0
repl = (inpt_df.PAY_6 == -2) | (inpt_df.PAY_6 == -1) | (inpt_df.PAY_6 == 0)
inpt_df.loc[repl, 'PAY_6'] = 0

# Saving cleansed file in csv format for part 2 - EDA

inpt_df.to_csv("inpt_df_processed.csv", index=None)
