==> 4. Data_Balancing.ipynb <==
## Class Balancing techniques application

In supervised learning, a common strategy to overcome the class imbalance problem is to resample the original training dataset to decrease the overall level of class imbalance. Resampling is done either by oversampling the minority (positive) class and/or under-sampling the majority (negative) class until the classes are approximately equally represented.

All resampling operations are applied to only training datasets. If upsampling is done before splitting the dataset into a train and validation set, then it could end up with the same observation in both datasets. As a result, a machine learning model will be able to perfectly predict the value for those observations when predicting on the validation set, hence inflating the accuracy and recall.

* Input : Transformed datasets from Notebook3 - x_train_outfix.csv, x_test_outfix.csv, y_train_c3.csv and y_test_c3.csv
* Outcome : Smote+TomekLink class balanced datasets - x_train_smtom.csv, y_train_smtom.csv, x_test_c4.csv and y_test_c4.csv

from __future__ import print_function 
import time

# Import libraries
import pandas as pd
import numpy as np
from pandas import DataFrame
import matplotlib.pyplot as plt
import seaborn as sns
import random
import sklearn
import scipy

import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

#Import Data balancing libraries
from imblearn.under_sampling import RandomUnderSampler, NearMiss
from imblearn.over_sampling import RandomOverSampler, SMOTE, ADASYN
from imblearn.combine import SMOTETomek, SMOTEENN
from collections import Counter

# Import models from sklearn
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier

# Import evaluation metrics
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
from sklearn.metrics import confusion_matrix, classification_report
from sklearn.metrics import roc_auc_score,roc_curve, auc
from sklearn.metrics import make_scorer
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.model_selection import cross_validate
from sklearn.metrics import cohen_kappa_score
from sklearn import model_selection

x_train=pd.read_csv('x_train_outfix.csv')
x_test=pd.read_csv('x_test_outfix.csv')

y_train=pd.read_csv('y_train_c3.csv')
y_test=pd.read_csv('y_test_c3.csv')

print(x_train.shape,x_test.shape,y_train.shape,y_test.shape)

### Functions for Model performance comparison

classifier = [
    ensemble.AdaBoostClassifier(), ensemble.BaggingClassifier(), XGBClassifier(),
    ensemble.GradientBoostingClassifier(), ensemble.RandomForestClassifier(), tree.DecisionTreeClassifier(),
    linear_model.LogisticRegressionCV(), naive_bayes.GaussianNB(), neighbors.KNeighborsClassifier(),svm.SVC(probability=True)
    ]

def models_comparison(x_train, y_train, x_test, y_test, folds):
    
    time_start = time.time()
    classifier_columns = []
    classifier_compare = pd.DataFrame(columns = classifier_columns)

    row_index = 0
    for alg in classifier:
    
        pred = alg.fit(x_train, y_train).predict(x_test)
        classifier_name = alg.__class__.__name__
        
        classifier_compare.loc[row_index,'ML Algorithm'] = classifier_name
        classifier_compare.loc[row_index, 'Train Accuracy'] = model_selection.cross_val_score(alg,x_train,y_train,cv=folds,scoring='accuracy').mean()
        classifier_compare.loc[row_index, 'Test Accuracy'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='accuracy').mean()
        classifier_compare.loc[row_index, 'Precision'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='precision').mean()
        classifier_compare.loc[row_index, 'Recall'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='recall').mean()
        classifier_compare.loc[row_index, 'F1 score'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='f1').mean()
        fp, tp, th = roc_curve(y_test, pred)
        classifier_compare.loc[row_index, 'ROC AUC'] = auc(fp, tp)       
        classifier_compare.loc[row_index, 'Kappa'] = cohen_kappa_score(y_test, pred, labels=None, weights=None, sample_weight=None)  
        roc_auc = auc(fp, tp)
        classifier_compare.loc[row_index, 'GINI'] = (2 * roc_auc) - 1
        tn, fp, fn, tp = confusion_matrix(y_test, pred, labels=[0,1]).ravel()
        classifier_compare.loc[row_index, 'Type II error'] = fn
        row_index+=1
    
    classifier_compare.sort_values(by = ['Test Accuracy'], ascending = False, inplace = True)  
    print('Time elapsed: {} seconds'.format(time.time()-time_start))
    return classifier_compare

### Under Sampling:

#### 1. Random undersampling

The simplest form of undersampling is to remove random records from the majority class. With imblearn's implementation we can choose to remove samples with or without replacement. The biggest drawback to this form of undersampling is loss of information.

rus =  RandomUnderSampler()
x_train_rus, y_train_rus = rus.fit_sample(x_train, y_train)
print(x_train_rus.shape,y_train_rus.shape)
print('\n',y_train_rus['DEFAULT'].value_counts())

#### 2. Near Miss

nm = NearMiss()
x_train_nm,y_train_nm = nm.fit_sample(x_train, y_train)
print(x_train_nm.shape,y_train_nm.shape)
print('\n',y_train_nm['DEFAULT'].value_counts())

### Over Sampling:

#### 3. Random oversampling

The simplest implementation of oversampling is to duplicate random records from the minority class, this can cause overfitting.

ros =  RandomOverSampler()
x_train_ros, y_train_ros = ros.fit_sample(x_train, y_train)
print(x_train_ros.shape,y_train_ros.shape)
print('\n',y_train_ros['DEFAULT'].value_counts())

#### 4. SMOTE (Synthetic Minority Over-Sampling Technique)

SMOTE synthesizes new examples by interpolating existing observations. SMOTE begins by iterating over every minority class instace and choosing its k nearest neighbors. The algorithm then constructs new instances halfway between the chosen obervations and its k neighbors. The greatest limitation of SMOTE is that it can only construct examples within the body of observations, never outside.

# Smote oversampling
sm = SMOTE()
x_train_smote, y_train_smote = sm.fit_sample(x_train, y_train)
print(x_train_smote.shape,y_train_smote.shape)
print('\n',y_train_smote['DEFAULT'].value_counts())

#### 5. ADASYN (Adaptive Synthetic Sampling)

ADASYN adaptively generates samples next to original observations which are wrongly classified by a KNN classifier. Unlike SMOTE that generates new samples that lie inside the class boundary, ADASYN tends to generate new samples near existing outliers.

adasyn =  ADASYN()
x_train_adasyn, y_train_adasyn = adasyn.fit_sample(x_train, y_train)
print(x_train_adasyn.shape,y_train_adasyn.shape)
print('\n',y_train_adasyn['DEFAULT'].value_counts())

### Hybrid Sampling:

#### 6.SMOTE+ENN

SMOTEENN is the combination of SMOTE and Edited Nearest Neighbor. ENN removes any example whose class label differs from the class label of at least two of its three nearest neighbors. ENN tends to remove more examples than the Tomek links.

smtenn =  SMOTEENN()
x_train_smtenn, y_train_smtenn = smtenn.fit_sample(x_train, y_train)
print(x_train_smtenn.shape,y_train_smtenn.shape)
print('\n',y_train_smtenn['DEFAULT'].value_counts())

#### 7. SMOTE+Tomek link

SMOTETomek is the combination of using Tomek links to undersample the majoirty class and the use of SMOTE to oversample the minority class.

smtom =  SMOTETomek()
x_train_smtom, y_train_smtom = smtom.fit_sample(x_train, y_train)
print(x_train_smtom.shape,y_train_smtom.shape)
print('\n',y_train_smtom['DEFAULT'].value_counts())

#### 8. Cost-sensitive learning

Cost-sensitive learning for imbalanced classification is focused on first assigning different costs to the types of misclassification errors that can be made, then using specialized methods to take those costs into account.
Instead of resampling with a focus on balancing the skewed class distribution, the focus is on changing the composition of the training dataset to meet the expectations of the cost matrix.

## class_weight='balanced' option will be assigned to algorithms directlyin the below section while running 
    # model comparison codes.The reason for its less popularity is that there are no cost-sensitive implementations 
    # of all learning algorithms.

### Selection of Sampling method based on ML model performance

10 Machine learning algorithms are considered with 10 fold cross validation

#1 No Sampling 

models_comparison(x_train, y_train, x_test, y_test, 10)

#2 Random UnderSampling

models_comparison(x_train_rus, y_train_rus, x_test, y_test, 10)

#3 NearMiss UnderSampling

models_comparison(x_train_nm, y_train_nm, x_test, y_test, 10)

#4 Random OverSampling

models_comparison(x_train_ros, y_train_ros, x_test, y_test, 10)

#5 SMOTE

models_comparison(x_train_smote, y_train_smote, x_test, y_test, 10)

#6 ADASYN

models_comparison(x_train_adasyn, y_train_adasyn, x_test, y_test, 10)

#7 SMOTE + ENN

models_comparison(x_train_smtenn, y_train_smtenn, x_test, y_test, 10)

#8 SMOTE + Tomeklink

models_comparison(x_train_smtom, y_train_smtom, x_test, y_test, 10)

#9 Cost sensitive learning

# Class Weights are used to correct class imbalances as a proxy for over \ undersampling. 
# Only below algorithms support cost-sensitive learning 

classifier = [
    XGBClassifier(scale_pos_weight=4), ensemble.RandomForestClassifier(class_weight='balanced'), 
    DecisionTreeClassifier(class_weight='balanced'), LogisticRegression(solver='lbfgs', class_weight='balanced')
    ]

models_comparison(x_train, y_train, x_test, y_test, 10)

### Conclusion

## Choosing best Sampling technique and why:
    
#SMOTE, ADASYN and SMOTE+TOMEK perform better than all other techniques
    #Train AccuracY, Test AccuracY, Precision, Recall AND F1 score metrics are similar among SMOTE, ADASYN and SMOTE+TOMEK
    # Metrics like  AUC, Kappa, GINI and Type II error, which are more efficient for evaluating imbalanced datasets are slightly
    # better for SMOTE+TOMEKLink method and also this method is least explored among researches. Hence this method is finalized.

# # Saving processed csv file for part 5 and 6 - Feature Selection and Dimentionality reduction

x_train_smtom.to_csv("x_train_smtom.csv", index=None)
y_train_smtom.to_csv("y_train_smtom.csv", index=None)

x_test.to_csv("x_test_c4.csv", index=None)
y_test.to_csv("y_test_c4.csv", index=None)
