==> 5. Feature Selection.ipynb <==
### Feature Selection

The input dataset consists of 24 independent variables.

The data features used to train machine learning models have a huge influence on mode performance. The irrelevant or partially relevant features can negatively impact model performance. Hence Feature selection methods are employed in this project to Reduces Overfitting, Improves Accuracy and Reduces Training Time.

* Input : Smote+TomekLink class balanced datasets - x_train_smtom.csv, y_train_smtom.csv, x_test_c4.csv and y_test_c4.csv
* Outcome : Selected Significant featues -
  fs_set3= ['LIMIT_BAL','PAY_AMT1', 'PAY_1', 'PAY_AMT2', 'AGE', 'SEX_2','PAY_3', 'MARRIAGE_2', 'BILL_AMT1',   'MARRIAGE_1', 'PAY_2', 'SEX_1', 'EDUCATION_2', 'EDUCATION_1']

 # Import libraries
from __future__ import print_function 
import time

# Import libraries
import pandas as pd
import numpy as np
from pandas import DataFrame
import matplotlib.pyplot as plt
import seaborn as sns
import random
import sklearn
import scipy
from collections import Counter

import warnings
warnings.filterwarnings('ignore')
%matplotlib inline

# Import feature selection libraries
from sklearn.feature_selection import mutual_info_classif, mutual_info_regression
from sklearn.feature_selection import RFE
from sklearn.feature_selection import chi2, SelectKBest, f_classif
from mlxtend.feature_selection import SequentialFeatureSelector as SFS
from mlxtend.feature_selection import ExhaustiveFeatureSelector as EFS
from mlxtend.plotting import plot_sequential_feature_selection as plot_sfs

# Import models from sklearn
from sklearn.model_selection import train_test_split, RandomizedSearchCV
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.svm import SVC
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import GradientBoostingClassifier
from xgboost import XGBClassifier

# Import metrics
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
from sklearn.metrics import confusion_matrix, classification_report
from sklearn.metrics import roc_auc_score,roc_curve, auc
from sklearn.metrics import make_scorer
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import f1_score
from sklearn.model_selection import cross_validate
from sklearn.metrics import cohen_kappa_score
from sklearn import model_selection

x_train=pd.read_csv('x_train_smtom.csv')
x_test=pd.read_csv('x_test_c4.csv')

y_train=pd.read_csv('y_train_smtom.csv')
y_test=pd.read_csv('y_test_c4.csv')

print(x_train.shape,x_test.shape,y_train.shape,y_test.shape)

### Filter Based methods:

#### 1. ANOVA Method

anova = SelectKBest(f_classif, k=12).fit(x_train, y_train)
print('Significant features from ANOVA method: {}'.format(x_train.columns[anova.get_support()]))

#### 2. Information Gain/Mutual Information

# calculate the mutual information between the variables and the target
# this returns the mutual information value of each feature
# the smaller the value the less information the feature has about the target

mutual_info = SelectKBest(mutual_info_classif, k=12).fit(x_train.fillna(0), y_train)
print('Significant features from Mutual Information method: {}'.format(x_train.columns[mutual_info.get_support()]))

### Wrapper based methods:

#### 3. Forward selection

sfs = SFS(RandomForestClassifier(), k_features=(5,12), forward=True, 
           floating=False, verbose=2,scoring='roc_auc',cv=0)
sfs = sfs.fit(x_train, y_train)

print('Significant features from Forward Selection method: {}'.format(x_train.columns[list(sfs.k_feature_idx_)]))

#### 5. Backward elimination

sbs = SFS(RandomForestClassifier(n_jobs=4), k_features=(5,12), forward=False, 
           floating=False, verbose=2, scoring='roc_auc', cv=0)
sbs = sbs.fit(x_train, y_train)

print('Significant features from Backward Elimination method: {}'.format(x_train.columns[list(sbs.k_feature_idx_)]))

#### 6. Stepwise method

sws = SFS(RandomForestClassifier(n_jobs=4), k_features=(5,12), forward=True, 
           floating=True, verbose=2,scoring='roc_auc', cv=0)
sws = sws.fit(np.array(x_train.fillna(0)), y_train)

print('Significant features from Stepwise selection method: {}'.format(x_train.columns[list(sws.k_feature_idx_)]))

#### 7. Recursive Feature Elimination using Random forests

# RFE will remove one least imp feature at each iteration by running random forest till a criteria is met.

rfe = RFE(RandomForestClassifier(), n_features_to_select=12)
rfe.fit(x_train.fillna(0), y_train)
print('Significant features from RFE method: {}'.format(x_train.columns[(rfe.get_support())]))

#### Shortlisting features based on their significance accross different feature selection methods

# 1 full dataset + 3 subset combinations

fs_set1=['LIMIT_BAL','PAY_AMT1', 'PAY_1', 'PAY_AMT2', 'AGE', 'SEX_2',
         'PAY_AMT4','BILL_AMT6', 'BILL_AMT1', 'PAY_AMT3', 'PAY_2', 'EDUCATION_2', 'MARRIAGE_2']

fs_set2=['LIMIT_BAL','PAY_AMT1', 'PAY_1', 'PAY_AMT2', 'AGE', 'SEX_1',
         'PAY_5','PAY_AMT3', 'BILL_AMT1', 'PAY_4', 'PAY_2', 'EDUCATION_1','MARRIAGE_1']

fs_set3=['LIMIT_BAL','PAY_AMT1', 'PAY_1', 'PAY_AMT2', 'AGE', 'SEX_2',
         'PAY_3', 'MARRIAGE_2', 'BILL_AMT1', 'MARRIAGE_1', 'PAY_2', 'SEX_1', 'EDUCATION_2', 'EDUCATION_1']

#### Functions to compare performace of ML models

classifier = [
    ensemble.AdaBoostClassifier(), ensemble.BaggingClassifier(), XGBClassifier(),
    ensemble.GradientBoostingClassifier(), ensemble.RandomForestClassifier(), tree.DecisionTreeClassifier(),
    linear_model.LogisticRegressionCV(), naive_bayes.GaussianNB(), neighbors.KNeighborsClassifier(),svm.SVC(probability=True)
    ]

def models_comparison(x_train, y_train, x_test, y_test, folds):
    
    time_start = time.time()
    classifier_columns = []
    classifier_compare = pd.DataFrame(columns = classifier_columns)

    row_index = 0
    for alg in classifier:
    
        pred = alg.fit(x_train, y_train).predict(x_test)
        classifier_name = alg.__class__.__name__
        
        classifier_compare.loc[row_index,'ML Algorithm'] = classifier_name
        classifier_compare.loc[row_index, 'Train Accuracy'] = model_selection.cross_val_score(alg,x_train,y_train,cv=folds,scoring='accuracy').mean()
        classifier_compare.loc[row_index, 'Test Accuracy'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='accuracy').mean()
        classifier_compare.loc[row_index, 'Precision'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='precision').mean()
        classifier_compare.loc[row_index, 'Recall'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='recall').mean()
        classifier_compare.loc[row_index, 'F1 score'] = model_selection.cross_val_score(alg,x_test,y_test,cv=folds,scoring='f1').mean()
        fp, tp, th = roc_curve(y_test, pred)
        classifier_compare.loc[row_index, 'ROC AUC'] = auc(fp, tp)       
        classifier_compare.loc[row_index, 'Kappa'] = cohen_kappa_score(y_test, pred, labels=None, weights=None, sample_weight=None)  
        roc_auc = auc(fp, tp)
        classifier_compare.loc[row_index, 'GINI'] = (2 * roc_auc) - 1
        tn, fp, fn, tp = confusion_matrix(y_test, pred, labels=[0,1]).ravel()
        classifier_compare.loc[row_index, 'Type II error'] = fn
        row_index+=1
    
    classifier_compare.sort_values(by = ['Test Accuracy'], ascending = False, inplace = True)  
    print('Time elapsed: {} seconds'.format(time.time()-time_start))
    return classifier_compare

### Selecting best set of features for modelling

#1. SET1 - fs_set1

models_comparison(x_train[fs_set1], y_train, x_test[fs_set1], y_test, 10)

#2. SET2 - fs_set2

models_comparison(x_train[fs_set2], y_train, x_test[fs_set2], y_test, 10)

#3. SET3 - fs_set3

models_comparison(x_train[fs_set3], y_train, x_test[fs_set3], y_test, 10)

### Conclusion

# Based on the performace of machine learning models, fs_set3 shows best results, especially for tree based models. 
    # Hence SET3 is finalized for further processing
# The models such as Logistic regression, SVC, KNN and Naive Bayes show slight improvement after feature selection, however 
    # it is not very significant
    
# Selected Significant features are --> 'LIMIT_BAL','PAY_AMT1', 'PAY_1', 'PAY_AMT2', 'AGE', 'SEX_2',
#          'PAY_3', 'MARRIAGE_2', 'BILL_AMT1', 'MARRIAGE_1', 'PAY_2', 'SEX_1', 'EDUCATION_2', 'EDUCATION_1'
